backend/pyproject.toml
[project]
name = "gettranscripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "databases[sqlite]>=0.9.0",
    "fastapi>=0.115.12",
    "google-api-python-client>=2.166.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "google-generativeai>=0.8.4",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.2",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.0",
    "youtube-transcript-api>=1.0.3",
]
